Add launchfile and .yaml-files for storing robot-specific params
* Added a launchfile for launching a controller with basic functionality
('twist_to_motors', 'pid_velocity'). Robot-specific parameters are
provided via .yaml-files. Parameter files for the gopigo3 and knex robots
are provided with this commit. Obviously, more parameter files can be
added when using the controller with other robots. The proper parameter
files for a robot are chosen by defining the environment variable 'ROBOT'
before launching the script 'differential_drive_yaml.launch'.

* This method should enable the package user to conveniently choose and
switch between parameters for a specific robot.
Parameter files for different robots can be collected in the repository
as they are of general use.